# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Dokumentasi API Final Project BE",
    "version": "1.0.0",
    "title": "Dokumentasi API Final Project BE",
    "contact": {
      "url": "https://www.github.com/"
    }
  },
  "host": "be-kel1.herokuapp.com",
...
  "schemes": ["https"],
  "paths": {
    "/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login user",
        "description": "",
        "operationId": "postLogin",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "form input",
            "required": true,
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil Login"
          },
          "400": {
            "description": "Gagal Login"
          }
        }
      }
    },
    "/register":{
      "post": {
        "tags": ["Auth"],
        "summary": "POST /register",
        "description": "Post register page",
        "operationId": "postRegister",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "form input",
            "required": true,
            "schema": {
              "$ref": "#/definitions/register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "GET Register success"
          }
        }
      }
    },
...
    "/product/user/sold": {
      "get": {
        "tags": ["Product"],
        "summary": "GET Product sold",
        "description": "",
        "operationId": "getProductSold",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [
          {
            "Authorization": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get Product sold"
          },
          "400": {
            "description": "Cannot get product sold"
          }
        }
      }
    },
    "/product": {
      "post": {
        "tags": ["Product"],
        "summary": "POST product",
        "description": "",
        "operationId": "postProduct",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "security": [
          {
            "Authorization": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in" : "formData",
            "description": "input name product",
            "required" : true,
            "type" : "string"
          },
          {
            "name": "category",
            "in" : "formData",
            "description": "input category product",
            "required" : true,
            "type" : "string",
            "enum": ["A","B","C","D"]
          },
          {
            "name": "price",
            "in" : "formData",
            "description": "input price product",
            "required" : true,
            "type" : "integer"
          },
          {
            "name": "description",
            "in" : "formData",
            "description": "input description product",
            "required" : true,
            "type" : "string"
          },
          {
            "name": "image",
            "in": "formData",
            "description": "input image product",
            "required": true,
            "type": "array",
            "items": {
              "type": "file",
              "format": "binary"
            }
          },
          {
            "name": "publish",
            "in": "formData",
            "description": "input image product",
            "required": true,
            "type": "boolean",
            "enum": [true, false]
          }
        ],
        "responses": {
          "200": {
            "description": "Get product"
          },
          "400": {
            "description": "Cannot get product"
          }
        }
      }
    },
...
    "/search/product?search={search}": {
      "get" : {
        "tags" : ["Product"],
        "summary" : "GET Search Product",
        "description" : "",
        "operationId" : "getSearchProduct",
        "consumes" : ["application/json"],
        "produces" : ["application/json"],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Get Search Product",
            "required": true,
            "type": "string"
          }
        ],
        "responses" : {
          "200" : {
            "description" : "Get Search Product Berhasil"
          },
          "400" : {
            "description" : "Get Search Product Gagal"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "login": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "test@test.com"
        },
        "password": {
          "type": "string",
          "example": "test"
        }
      }
    },
    "register": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Budiman"
        },
        "email": {
          "type": "string",
          "example": "budiman@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "budiman"
        }
      }
    },
...
  },
...
}